+++ ts/features/bonus/bonusVacanze/components/ActiveBonus.tsx
@@ -1,15 +1,15 @@
 import * as O from "fp-ts/lib/Option";
-import { Text, View } from "native-base";
+import { Text as NBText, View } from "native-base";
 import * as React from "react";
 import { StyleSheet } from "react-native";
 import { BonusActivationWithQrCode } from "../../../../../definitions/bonus_vacanze/BonusActivationWithQrCode";
+import { IOColors } from "../../../../components/core/variables/IOColors";
 import TouchableDefaultOpacity from "../../../../components/TouchableDefaultOpacity";
 import IconFont from "../../../../components/ui/IconFont";
 import I18n from "../../../../i18n";
 import customVariables from "../../../../theme/variables";
 import { formatNumberAmount } from "../../../../utils/stringBuilder";
 import { validityInterval } from "../utils/bonus";
-import { IOColors } from "../../../../components/core/variables/IOColors";
 
 type Props = {
   bonus: BonusActivationWithQrCode;
@@ -85,30 +85,30 @@ const ActiveBonus: React.FunctionComponent<Props> = (props: Props) => {
     >
       <View style={styles.spaced}>
         {O.isSome(bonusValidityInterval) && (
-          <Text>{`${I18n.t("bonus.bonusVacanze.validity")} ${
+          <NBText>{`${I18n.t("bonus.bonusVacanze.validity")} ${
             bonusValidityInterval.value.e1
-          } - ${bonusValidityInterval.value.e2}`}</Text>
+          } - ${bonusValidityInterval.value.e2}`}</NBText>
         )}
-        <Text bold={true} style={styles.text12}>
+        <NBText bold={true} style={styles.text12}>
           {formatNumberAmount(props.bonus.dsu_request.max_amount, true)}
-        </Text>
+        </NBText>
       </View>
       <View small={true} />
       <View style={styles.spaced}>
-        <Text style={styles.textWhite}>
+        <NBText style={styles.textWhite}>
           {I18n.t("bonus.bonusVacanze.taxBenefit")}
-        </Text>
-        <Text bold={true} style={styles.text12}>
+        </NBText>
+        <NBText bold={true} style={styles.text12}>
           {formatNumberAmount(props.bonus.dsu_request.max_tax_benefit, true)}
-        </Text>
+        </NBText>
       </View>
       <View style={styles.smallSpacer} />
       <View style={styles.text3Line}>
         <View style={styles.text3Container}>
-          <Text numberOfLines={2} style={styles.text3}>
+          <NBText numberOfLines={2} style={styles.text3}>
             {/* TODO replace this hardcoded string */}
             {"Bonus Vacanze"}
-          </Text>
+          </NBText>
         </View>
         <View style={styles.icon}>
           <IconFont name="io-right" size={ICON_WIDTH} color={IOColors.white} />