+++ ts/sagas/__tests__/initializeApplicationSaga.test.ts
@@ -76,7 +76,6 @@ describe("initializeApplicationSaga", () => {
       .fork(watchSessionExpiredSaga)
       .next()
       .next(200) // checkSession
-      .next() // updateInstallationSaga
       .select(sessionInfoSelector)
       .next(O.none)
       .next(O.none) // loadSessionInformationSaga
@@ -140,7 +139,6 @@ describe("initializeApplicationSaga", () => {
       .fork(watchSessionExpiredSaga)
       .next()
       .next(200) // check session
-      .next() // updateInstallationSaga
       .select(sessionInfoSelector)
       .next(
         O.some({