+++ ts/sagas/startup.ts
@@ -27,12 +27,14 @@ import {
   bpdEnabled,
   cdcEnabled,
   euCovidCertificateEnabled,
+  fciEnabled,
   mvlEnabled,
   pagoPaApiUrlPrefix,
   pagoPaApiUrlPrefixTest,
   pnEnabled,
   svEnabled,
-  zendeskEnabled
+  zendeskEnabled,
+  idPayEnabled
 } from "../config";
 import { watchBonusSaga } from "../features/bonus/bonusVacanze/store/sagas/bonusSaga";
 import { watchBonusBpdSaga } from "../features/bonus/bpd/saga";
@@ -41,6 +43,7 @@ import { watchBonusSvSaga } from "../features/bonus/siciliaVola/saga";
 import { watchEUCovidCertificateSaga } from "../features/euCovidCert/saga";
 import { watchMvlSaga } from "../features/mvl/saga";
 import { watchZendeskSupportSaga } from "../features/zendesk/saga";
+import { watchFciSaga } from "../features/fci/saga";
 import I18n from "../i18n";
 import { mixpanelTrack } from "../mixpanel";
 import NavigationService from "../navigation/NavigationService";
@@ -78,6 +81,7 @@ import { differentProfileLoggedIn } from "../store/actions/crossSessions";
 import { clearAllMvlAttachments } from "../features/mvl/saga/mvlAttachments";
 import { watchMessageAttachmentsSaga } from "../features/messages/saga/attachments";
 import { watchPnSaga } from "../features/pn/store/sagas/watchPnSaga";
+import { watchIDPayWalletSaga } from "../features/idpay/wallet/saga";
 import {
   startAndReturnIdentificationResult,
   watchIdentification
@@ -222,10 +226,6 @@ export function* initializeApplicationSaga(): Generator<
     yield* put(sessionExpired());
     return;
   }
-
-  // Start the notification installation update as early as
-  // possible to begin receiving push notifications
-  yield* call(updateInstallationSaga, backendClient.createOrUpdateInstallation);
 
   // whether we asked the user to login again
   const isSessionRefreshed = previousSessionToken !== sessionToken;
@@ -378,6 +378,10 @@ export function* initializeApplicationSaga(): Generator<
     }
   }
 
+  // Start the notification installation update as early as
+  // possible to begin receiving push notifications
+  yield* call(updateInstallationSaga, backendClient.createOrUpdateInstallation);
+
   //
   // User is autenticated, session token is valid
   //
@@ -423,6 +427,15 @@ export function* initializeApplicationSaga(): Generator<
   if (mvlEnabled || pnEnabled) {
     // Start watching for message attachments actions
     yield* fork(watchMessageAttachmentsSaga, sessionToken);
+  }
+
+  if (idPayEnabled) {
+    // Start watching for IDPay wallet actions
+    yield* fork(watchIDPayWalletSaga, sessionToken);
+  }
+
+  if (fciEnabled) {
+    yield* fork(watchFciSaga, sessionToken);
   }
 
   // Load the user metadata